refactor(coverage): modularize
